:hammer: Add type parameter to divider

+++ apu/divider_test.go
@@ -10,7 +10,7 @@ func Test_divider(t *testing.T) {
 	ctx, cancel := context.WithTimeout(context.Background(), 3*time.Second)
 	defer cancel()
 
-	d := runDivider(ctx, 5)
+	d := runDivider(ctx, uint(5))
 	ch := d.output()
 
 	for i := 0; i < 5; i++ {
